requirements.txt
# Web Framework
Flask==3.1.0
flask-cors==5.0.1
flask-openapi3[swagger]==4.1.0

# Database
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2

# Data Validation
pydantic==2.10.6

# Security & Encryption
cryptography==45.0.5

# Data Processing
pandas==2.3.0
numpy==2.2.6

# Machine Learning
scikit-learn==1.7.0
joblib==1.5.1

# Deep Learning & NLP
torch==2.7.1
transformers==4.52.4
datasets==3.6.0

# Google AI
google-genai==1.53.0

# Progress Bars
tqdm==4.67.1

# Testing
pytest==8.4.0

# Environment Variables
python-dotenv==1.2.1

# Production Server
gunicorn==21.2.0

# Kafka
quixstreams==2.4